Rename: index.tsx 컴포넌트명으로 변경
