+++ src/pages/Theme/ThemePage.tsx
@@ -1,0 +1,37 @@
+import axios from 'axios';
+import { useEffect } from 'react';
+import { useState } from 'react';
+import { useParams } from 'react-router-dom';
+
+import { ThemeGoodsSection } from '@/components/features/Theme/ThemeGoodsSection/ThemeGoodsSection';
+import { ThemeHeroSection } from '@/components/features/Theme/ThemeHeroSection/ThemeHeroSection';
+import type { ThemeData } from '@/types/types';
+
+export const ThemePage = () => {
+  const { themeKey = '' } = useParams<{ themeKey: string }>();
+  const [themes, setThemes] = useState<ThemeData[]>([]);
+
+  useEffect(() => {
+    const fetchThemes = async () => {
+      try {
+        const response = await axios.get(
+          'https://react-gift-mock-api-ten.vercel.app/api/v1/themes',
+        );
+        setThemes(response.data.themes);
+      } catch (error) {
+        console.error('Error fetching themes:', error);
+      }
+    };
+
+    fetchThemes();
+  }, []);
+
+  console.log(themes);
+
+  return (
+    <>
+      <ThemeHeroSection themes={themes} themeKey={themeKey} />
+      <ThemeGoodsSection themeKey={themeKey} />
+    </>
+  );
+};
