--- src/pages/Theme/index.tsx
@@ -1,43 +1,0 @@
-import axios from 'axios';
-import { useEffect, useState } from 'react';
-import { Navigate, useParams } from 'react-router-dom';
-
-import { ThemeGoodsSection } from '@/components/features/Theme/ThemeGoodsSection';
-import { ThemeHeroSection } from '@/components/features/Theme/ThemeHeroSection';
-import { getCurrentTheme } from '@/components/features/Theme/ThemeHeroSection';
-import { RouterPath } from '@/routes/path';
-import type { ThemeData } from '@/types';
-
-export const ThemePage = () => {
-  const { themeKey = '' } = useParams<{ themeKey: string }>();
-  const [Themes, setThemes] = useState<ThemeData[]>([]);
-
-  useEffect(() => {
-    const fetchThemes = async () => {
-      try {
-        const response = await axios.get(
-          'https://react-gift-mock-api-ten.vercel.app/api/v1/themes',
-        );
-        setThemes(response.data.themes);
-      } catch (error) {
-        console.error('Error fetching themes:', error);
-      }
-    };
-
-    fetchThemes();
-  }, []);
-
-  const currentTheme = getCurrentTheme(themeKey, Themes);
-  console.log(currentTheme);
-
-  if (!currentTheme) {
-    return <Navigate to={RouterPath.notFound} />;
-  }
-
-  return (
-    <>
-      <ThemeHeroSection themeKey={themeKey} />
-      <ThemeGoodsSection themeKey={themeKey} />
-    </>
-  );
-};
