+++ src/provider/Auth/Auth.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from 'react';
 import { createContext, useContext, useEffect, useState } from 'react';
 
-import { authSessionStorage } from '@/utils/storage';
+import { authSessionStorage } from '@/utils/storage/storage';
 
 type AuthInfo = {
   id: string;
